Tweaks to node update method
We override the update method so that the allowed_methods property can
be stripped before being sent to the server, avoiding validation issues.

We also introduce a new has_method method, which can be used in the UI
to check if the user can perform a specific method on the node.